Allow custom URL for Alloc
Because AllocCLI, by default, generates the Alloc URL from the Alloc
root directory, you must either store AllocCLI in a directory whose name
reflects the FQDN of the Alloc host, or pass the URL as an argument to
the Alloc object during initialization. This patch allows a custom url
to be passed via an ALLOC_URL environment variable.